Trails graphics added
Added the trails of the campus
