+++ index.html
@@ -201,7 +201,6 @@
 
 <!-- Link to external JavaScript file -->
 <script src="showBuildingPhoto.js"></script>
-
 </body>
 
 </html>
