+++ trailsData.js
@@ -1,0 +1,117 @@
+//Database file of all the trails and their coordinates
+//const trails  
+const trailsData = [  
+    {
+    description: "dormsToEntranceTrail", 
+    x1: 3871044.48,
+    y1: 3765127.84,
+    x2: 3870932.69,
+    y2: 3765134.02
+    },
+    { 
+    description: "mainEntranceTrail",
+    x1: 3870934.28,
+    y1: 3765134.02,
+    x2: 3870919.31,
+    y2: 3765126.29
+    },
+    {
+    description: "dormToBuilding6", 
+    x1: 3871044.48,
+    y1: 3765127.84,
+    x2: 3870994.86,
+    y2: 3765211.47
+    },
+    {
+      description: "building8trail",
+      x1: 3870937.94,
+      y1: 3765135.85,
+      x2: 3870944.69,
+      y2: 3765183.33
+    },
+    {
+      description: "building8trailLeftTurn",
+      x1: 3870944.69,
+      y1: 3765183.33,
+      x2: 3870913.45,
+      y2: 3765188.49
+    },
+    {
+      description: "building8ToBuilding5",
+      x1: 3870913.45,
+      y1: 3765188.49,
+      x2: 3870919.50,
+      y2: 3765219.34
+    },
+    {
+      description: "building5Left",
+      x1: 3870919.50,
+      y1: 3765219.34,
+      x2: 3870904.32,
+      y2: 3765238.76
+    },
+    {
+      description: "sculpturesTrail",
+      x1: 3870919.50,
+      y1: 3765219.34,
+      x2: 3870923.36,
+      y2: 3765352.78
+    },
+    {
+      description: "building3ToBuilding5",
+      x1: 3870921.15,
+      y1: 3765288.61,
+      x2: 3870973.68,
+      y2: 3765261.22
+    },
+    {
+      description: "building5ToBuilding6Meeting",
+      x1: 3870972.77,
+      y1: 3765262.28,
+      x2: 3871010.83,
+      y2: 3765232.48
+    },
+    {
+      description: "building6ToExit",
+      x1: 3870995.81,
+      y1: 3765210.87,
+      x2: 3871025.19,
+      y2: 3765251.84
+    },
+    {
+      description: "building8ToBuilding6",
+      x1: 3870944.15,
+      y1: 3765183.35,
+      x2: 3870996.31,
+      y2: 3765212.35
+    },
+    {
+      description: "rightTurnBuilding1",
+      x1: 3870923.17,
+      y1: 3765352.43,
+      x2: 3870945.13,
+      y2: 3765352.32
+    },
+    {
+      description: "leftTurnBuilding1",
+      x1: 3870923.17,
+      y1: 3765352.43,
+      x2: 3870881.63,
+      y2: 3765360.46
+    },
+    {
+      description: "leftTurnBuilding1",
+      x1: 3870882.86,
+      y1: 3765360.54,
+      x2: 3870885.20,
+      y2: 3765385.97
+    },
+    {
+      description: "leftTurnBuilding1",
+      x1: 3870943.92,
+      y1: 3765352.76,
+      x2: 3870944.58,
+      y2: 3765381.02
+    },
+    ];
+    export default trailsData;
